backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.1",
    "chromadb>=1.2.1",
    "httpx>=0.28.1",
    "ipython>=8.37.0",
    "langchain>=1.0.2",
    "langchain-community>=0.3.31",
    "langchain-openai>=1.0.1",
    "livekit-agents>=1.2.15",
    "livekit-plugins-deepgram>=1.2.15",
    "livekit-plugins-elevenlabs>=1.2.15",
    "livekit-plugins-openai>=1.2.15",
    "livekit-plugins-silero>=1.2.15",
    "llama-index>=0.14.5",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-vector-stores-pinecone>=0.7.1",
    "ormsgpack>=1.11.0",
    "pinecone-client>=6.0.0",
    "pydantic>=2.12.3",
    "pypdf>=6.1.2",
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
]
